More clearly define support
This started out as a plan to support Django 1.7 and 1.8, and on all
current versions of Python. This turned into supporting Django 1.8 on
Python 2.7, 3.3, and 3.4 only.

The tests don't pass under Django 1.7 because 1.8 introduced new
settings for configuring templates. Because of this the install_requires
section of setup.py declares that it must be using Django 1.8+.

Python 3.3 support was added, since that's a supported version of Python
under Django 1.8.

Also added pytest-xdist to the tox dependencies, to support running

    tox -e py34 -- -f

in order to have it watch and rerun the tests in development.